# chore: Refactor manifest generation E2E tests to unit tests