… e2e/tests/manifest-content.test.ts
… src/core/utils/__tests__/manifest.test.ts