Most these tests where just testing one function, so it's easier and faster to write them as unit tests instead of E2E tests.

| CI Job | Before | After |
| --- | :---: |  :---: | 
| `tests` | 1m 11s | 53s - 59s |
| `windows-tests` | 2m 33s | 1m 36s - 1m 58s |

> Windows tests are the slowest job in the validation workflow, so any improvements there effect the total time it takes to validate a commit.

This addresses the following TODO:

https://github.com/wxt-dev/wxt/blob/4f649fc8724a3c7f2fae158ad2ac5ecc495995cd/e2e/tests/manifest-content.test.ts#L4